Make ModifiedClassPathRunner somewhat compatible with Java 9
In Java 9, the system class loader is no longer a URLClassLoader so
we no longer have a direct route to the URLs that are on the class
path. We can, however, use reflection to similar effect which is
what this commit does. It's undoubtedly brittle and also relies on
opening up a portion of the java.base module:

--add-opens=java.base/jdk.internal.loader=ALL-UNNAMED

However, it works (with 9-ea+168 at least) and that's good enough
for now.